USB: gadget: ci13xx_udc: Fix bug in halting a control endpoint
A new setup packet can arrive on the control endpoint exactly when
software tries to halt it.  The halt condition will continue to be
cleared by the hardware.  The current code is ignoring this scenario
and polling for the halt condition to be true.  Return success from
ep_set_halt upon a new setup packet arrival.

CRs-fixed: 299838
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>