vidc: 720p: Memset MPEG-4 recon buffers to black
In the case of MPEG-4 decoding with clips which have
initial frames starting with P-VOPs, green macroblocks
are seen. This is because of recon buffers used for
MPEG-4 have 0 preset values. The default values are
changed to 0x10 & 0x80 for luma and chroma to get
default color as black.

CRs-Fixed: 303298
Signed-off-by: Rajeshwar Kurapaty <rkurapat@codeaurora.org>