tty: serial: msm_serial_hs: remove dma_wake_lock
For some reason we grab the dma_wake_lock as soon as the module is
initialized, preventing the power management runtime from suspending.
The only way to release the dma_wake_lock is through the device's
power management runtime_suspend function, meaning this lock is acquired
on device boot and released on shutdown only.